// src/features/users/schemas/update-user.schema.ts
import { z } from "zod";

export const updateUserSchema = z.object({
    email: z.string().email("Email inválido").optional(),
    username: z.string().min(3, "Username deve ter no mínimo 3 caracteres").optional(),
    password: z.string().min(6, "Senha deve ter no mínimo 6 caracteres").optional(),
    firstname: z.string().min(2, "Nome deve ter no mínimo 2 caracteres").optional(),
    lastname: z.string().min(2, "Sobrenome deve ter no mínimo 2 caracteres").optional(),
    city: z.string().min(2, "Cidade deve ter no mínimo 2 caracteres").optional(),
    street: z.string().min(3, "Rua deve ter no mínimo 3 caracteres").optional(),
    number: z.string().min(1, "Número é obrigatório").optional(),
    zipcode: z.string().min(5, "CEP deve ter no mínimo 5 caracteres").optional(),
    lat: z.string().optional(),
    long: z.string().optional(),
    phone: z.string().min(8, "Telefone deve ter no mínimo 8 caracteres").optional(),
});

export type UpdateUserFormData = z.infer<typeof updateUserSchema>;